Renamed timing_rp2040 to platform_timing
