+++ firmware/include/bmp/platform_timing.h
