+++ firmware/source/bmp/gdb_if.c
@@ -30,7 +30,7 @@
 
 #include "FreeRTOS.h"
 #include "task.h"
-#include "timing_rp2040.h"
+#include "platform_timing.h"
 
 #include "tusb.h"
